design #89 라이브 뱃지 추가
