+++ Projects/Features/MainFeature/Sources/PaddingUILabel.swift
@@ -1,0 +1,16 @@
+import UIKit
+
+final class PaddingUILabel: UILabel {
+    var textInsets = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
+    
+    override func drawText(in rect: CGRect) {
+        let insetRect = rect.inset(by: textInsets)
+        super.drawText(in: insetRect)
+    }
+
+    override var intrinsicContentSize: CGSize {
+        let size = super.intrinsicContentSize
+        return CGSize(width: size.width + textInsets.left + textInsets.right,
+                      height: size.height + textInsets.top + textInsets.bottom)
+    }
+}
